fix login funtion
