+++ components/front-service/package.json
@@ -14,10 +14,12 @@
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
     "@mui/icons-material": "^5.14.16",
-    "@mui/lab": "^6.0.0-beta.11",
+    "@mui/lab": "^5.0.0-alpha.173",
+    "@mui/material": "^5.16.7",
     "@mui/styled-engine-sc": "^6.0.0-alpha.4",
     "@mui/x-date-pickers": "^7.20.0",
     "axios": "^1.6.0",
+    "keycloak-js": "^26.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.17.0",
