+++ components/front-service/src/App.tsx
@@ -1,5 +1,6 @@
-import React from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import React, { useEffect, useState } from 'react';
+import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
+import keycloak from './Keycloak';
 import Admin from './pages/Admin';
 import AdminLogin from './pages/AdminLogin';
 import AdminCreateAccount from './pages/AdminCreateAccount';
@@ -19,25 +20,45 @@ import UserDetails from './pages/UserDetails';
 import './App.css';
 
 const App: React.FC = () => {
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [loading, setLoading] = useState(true);
+
+
+  useEffect(() => {
+    keycloak.init({ onLoad: 'login-required' })
+      .then(authenticated => {
+        setIsAuthenticated(authenticated);
+        setLoading(false);
+      })
+      .catch(error => {
+        console.error('Failed to initialize Keycloak:', error);
+        setLoading(false);
+      });
+  }, []);
+
+  if (loading) {
+    return <div>loggin in...</div>;
+  }
+
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Dashboard />} />
-        <Route path="/login" element={<Login />} />
-        <Route path="/user/:id" element={<UserDetails />} />
-        <Route path="/search" element={<SearchResultDashboard />} />
-        <Route path="/like/post" element={<FavoritePostList />} />
-        <Route path="/like/get" element={<FavoriteGetList />} />
-        <Route path="/chat/" element={<ChatList />} />
-        <Route path="/chat/:id" element={<Chat />} />
-        <Route path="/account/create" element={<AccountCreate />} />
-        <Route path="/account/create/p" element={<ProfileCreate />} />
-        <Route path="/account" element={<MyAccount />} />
         <Route path="/admin/login" element={<AdminLogin />} />
         <Route path="/admin" element={<Admin />} />
         <Route path="/admin/account/create" element={<AdminCreateAccount />} />
         <Route path="/admin/account/edit" element={<AdminEditAccount />} />
         <Route path="/admin/account/delete" element={<AdminDeleteAccount />} />
+        <Route path="/" element={isAuthenticated ? <Dashboard /> : <Navigate to="/" />} />
+        <Route path="/user/:id" element={<UserDetails />} />
+        <Route path="/search" element={<SearchResultDashboard />} />
+        <Route path="/like/post" element={<FavoritePostList />} />
+        <Route path="/like/get" element={<FavoriteGetList />} />
+        <Route path="/chat/" element={isAuthenticated ? <ChatList /> : <Navigate to="/login" />} />
+        <Route path="/chat/:id" element={<Chat />} />
+        <Route path="/account/create" element={<AccountCreate />} />
+        <Route path="/account/create/p" element={<ProfileCreate />} />
+        <Route path="/account" element={<MyAccount />} />
+        <Route path="/login" element={<Login />} />
       </Routes>
     </Router>
   );
